Refactor item microfactor rendering
