+++ src/components/item-microformat/index.js
@@ -2,12 +2,16 @@ import DocumentItem from './document-item';
 import ProductItem from './product-item';
 import FolderItem from './folder-item';
 
-export default function items(item) {
+export default function ItemMicroformat({ item }) {
+  if (!item) {
+    return null;
+  }
+
   const types = {
     product: <ProductItem data={item} key={item.path} />,
     folder: <FolderItem data={item} key={item.path} />,
     document: <DocumentItem data={item} key={item.path} />
   };
 
-  return types[item.type];
+  return types[item.type] || null;
 }
