+++ src/components/topics/index.js
@@ -1,4 +1,4 @@
-import items from 'components/items';
+import ItemMicroformat from 'components/item-microformat';
 
 import { TopicMap, List } from './styles';
 
@@ -7,9 +7,13 @@ export default function Topics({ topicMaps }) {
     return null;
   }
 
-  return topicMaps?.map(topic => (
+  return topicMaps?.map((topic) => (
     <TopicMap key={topic.name}>
-      <List>{topic.items.edges.map(({ node }) => items(node))}</List>
+      <List>
+        {topic.items.edges.map(({ node }, i) => (
+          <ItemMicroformat item={node} key={i} />
+        ))}
+      </List>
     </TopicMap>
   ));
 }
